# perf(reporters): update summary only when needed